refactor(logger-plugin): extract ActionLogger for logging action lifecycle
